Fix problem with resubmitting the new assignment form
On submit, if the client was already assigned to this attendance sheet,
the save would cancel.  However, if the user changed the client name and
resubmitted it would not reset the client to the newly selected client.
This change fixes that but introduces the possibility for the user to
edit an already saved assignment and save an already assigned client.

Signed-off-by: Ian Hilt <ihilt@eastway.org>